services/web/requirements.txt
flask
flask_limiter
flask_cors
python-dotenv
requests
gunicorn